Edit lines, update documentation

+++ R/Test Individual Variables.R
@@ -37,10 +37,10 @@ TestIndBRS <- function(fun, time_flags1, time_flags2, thr = 0.5,
 TestBRSDWT <- function(fun, time_flags1, time_flags2){
   time_flags1 <- time_flags1 * 60
   time_flags2 <- time_flags2 * 60
-  select_time1 <- fun$Time[(fun$Time >= as.integer(time_flags1[1])) &
-                              (fun$Time <= as.integer(time_flags1[2]))]
-  select_time2 <- fun$Time[(fun$Time >= as.integer(time_flags2[1])) &
-                             (fun$Time <= as.integer(time_flags2[2]))]
+  select_time1 <- fun$Time[(fun$Time >= as.numeric(time_flags1[1])) &
+                              (fun$Time <= as.numeric(time_flags1[2]))]
+  select_time2 <- fun$Time[(fun$Time >= as.numeric(time_flags2[1])) &
+                             (fun$Time <= as.numeric(time_flags2[2]))]
   select_time1 <- match(select_time1, fun$Time)
   select_time2 <- match(select_time2, fun$Time)
   if((sum(select_time1 %in% select_time2)) != 0){
@@ -80,10 +80,10 @@ TestBRSCWT <- function(fun, time_flags1, time_flags2, thr = 0.5, use.thr = TRUE)
   VLF <- fun$VLF
   time_flags1 <- time_flags1 * 60
   time_flags2 <- time_flags2 * 60
-  select_time1 <- fun$Time[(fun$Time >= as.integer(time_flags1[1])) &
-                             (fun$Time <= as.integer(time_flags1[2]))]
-  select_time2 <- fun$Time[(fun$Time >= as.integer(time_flags2[1])) &
-                             (fun$Time <= as.integer(time_flags2[2]))]
+  select_time1 <- fun$Time[(fun$Time >= as.numeric(time_flags1[1])) &
+                             (fun$Time <= as.numeric(time_flags1[2]))]
+  select_time2 <- fun$Time[(fun$Time >= as.numeric(time_flags2[1])) &
+                             (fun$Time <= as.numeric(time_flags2[2]))]
   select_time1 <- match(select_time1, fun$Time)
   select_time2 <- match(select_time2, fun$Time)
   if((sum(select_time1 %in% select_time2)) != 0){
@@ -156,12 +156,12 @@ TestIndHRandBP <- function(data, time_flags1, time_flags2){
   data$HR <- 60000/data$RR
   #data <- list(Time = data[,"Time"], HR = 60000/data[,"RR"], SBP = data[,"SBP"])
   time_flags1 <- time_flags1 * 60
-  select_time1 <- data$Time[(data$Time >= as.integer(time_flags1[1])) &
-                             (data$Time <= as.integer(time_flags1[2]))]
+  select_time1 <- data$Time[(data$Time >= as.numeric(time_flags1[1])) &
+                             (data$Time <= as.numeric(time_flags1[2]))]
   select_time1 <- match(select_time1, data$Time)
   time_flags2 <- time_flags2 * 60
-  select_time2 <- data$Time[(data$Time >= as.integer(time_flags2[1])) &
-                             (data$Time <= as.integer(time_flags2[2]))]
+  select_time2 <- data$Time[(data$Time >= as.numeric(time_flags2[1])) &
+                             (data$Time <= as.numeric(time_flags2[2]))]
   select_time2 <- match(select_time2, data$Time)
   if((sum(select_time1 %in% select_time2)) != 0){
     if(((min(select_time1) - min(select_time2)) < 0) & 
@@ -212,12 +212,12 @@ TestIndHRandBP <- function(data, time_flags1, time_flags2){
 #' TestIndHRV(AlphaIndex, c(0, 1.7), c(8, 9.5))
 TestIndHRV <- function(fun, time_flags1, time_flags2){
   time_flags1 <- time_flags1 * 60
-  select_time1 <- fun$Time[(fun$Time >= as.integer(time_flags1[1])) &
-                             (fun$Time <= as.integer(time_flags1[2]))]
+  select_time1 <- fun$Time[(fun$Time >= as.numeric(time_flags1[1])) &
+                             (fun$Time <= as.numeric(time_flags1[2]))]
   select_time1 <- match(select_time1, fun$Time)
   time_flags2 <- time_flags2 * 60
-  select_time2 <- fun$Time[(fun$Time >= as.integer(time_flags2[1])) &
-                             (fun$Time <= as.integer(time_flags2[2]))]
+  select_time2 <- fun$Time[(fun$Time >= as.numeric(time_flags2[1])) &
+                             (fun$Time <= as.numeric(time_flags2[2]))]
   select_time2 <- match(select_time2, fun$Time)
   if((sum(select_time1 %in% select_time2)) != 0){
     if(((min(select_time1) - min(select_time2)) < 0) &
